Refactor: fixtures 디렉터리의 파일들을 구조에 맞게 분류
