--- fixtures/fieldState.js
@@ -1,6 +1,0 @@
-const fieldState = {
-  value: '',
-  validationMessage: '',
-};
-
-export default fieldState;
