--- fixtures/footerMenuList.js
@@ -1,39 +1,0 @@
-const footerMenuList = [
-  {
-    id: 1,
-    column: 1,
-    rowStart: 1,
-    rowEnd: 2,
-    title: 'Home',
-  },
-  {
-    id: 2,
-    column: 1,
-    rowStart: 2,
-    rowEnd: 3,
-    title: 'About',
-  },
-  {
-    id: 3,
-    column: 1,
-    rowStart: 3,
-    rowEnd: 4,
-    title: 'Terms of use',
-  },
-  {
-    id: 4,
-    column: 1,
-    rowStart: 4,
-    rowEnd: 5,
-    title: 'Shopping',
-  },
-  {
-    id: 5,
-    column: 2,
-    rowStart: 1,
-    rowEnd: 2,
-    title: 'Privacy Policy',
-  },
-];
-
-export default footerMenuList;
