--- fixtures/initialLoginFields.js
@@ -1,21 +1,0 @@
-const initialLoginFields = {
-  email: {
-    value: '',
-    validationMessage: '',
-  },
-  password: {
-    value: '',
-    validationMessage: '',
-  },
-  firstName: {
-    value: '',
-    validationMessage: '',
-  },
-  lastName: {
-    value: '',
-    validationMessage: '',
-  },
-  error: '',
-};
-
-export default initialLoginFields;
