--- fixtures/inputList.js
@@ -1,8 +1,0 @@
-const inputList = {
-  lastName: 'Last Name은',
-  firstName: 'First Name은',
-  email: 'Email은',
-  password: 'Password는',
-};
-
-export default inputList;
