--- fixtures/listCategoriesStyle.js
@@ -1,23 +1,0 @@
-const listCategoriesStyle = {
-  container: {
-    margin: '2rem 0 0 2rem',
-    '& h1': {
-      fontSize: '20px',
-      fontWeight: '700',
-    },
-  },
-  button: {
-    all: 'unset',
-    display: 'block',
-    marginTop: '1.5rem',
-    fontSize: '17px',
-    fontWeight: '700',
-    color: '#696871',
-    cursor: 'pointer',
-    '&:hover': {
-      color: '#333332',
-    },
-  },
-};
-
-export default listCategoriesStyle;
