--- fixtures/loginFailData.js
@@ -1,7 +1,0 @@
-const loginFailData = {
-  error: {
-    message: 'INVALID_PASSWORD',
-  },
-};
-
-export default loginFailData;
