--- fixtures/loginTokens.js
@@ -1,6 +1,0 @@
-const loginTokens = {
-  refreshToken: 'refreshToken',
-  localId: 'localId',
-};
-
-export default loginTokens;
