--- fixtures/postLoginData.js
@@ -1,6 +1,0 @@
-const postLoginData = {
-  email: '',
-  password: '',
-};
-
-export default postLoginData;
