--- fixtures/signUpFailData.js
@@ -1,7 +1,0 @@
-const signUpFailData = {
-  error: {
-    message: 'ID_ALREADY_EXISTS',
-  },
-};
-
-export default signUpFailData;
