+++ src/Footer/FooterTopList.test.jsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 
 import FooterTopList from './FooterTopList';
 
-import FOOTER_MENU_LIST from '../../fixtures/footerMenuList';
+import FOOTER_MENU_LIST from '../../fixtures/Footer/footerMenuList';
 
 describe('FooterTopList', () => {
   it('renders the list', () => {
